Add eBPF loader and test program
BPF UPROBE program expanded to allow for injecting into the first
parameter of a function call. This parameter is expected to be a user
memory address.

Loader can load the BPF program described above with parameters passed
from the command lines.

Test program as an example to inject into with a probe.

Signed-off-by: Vito Tomas <vito.tomas205@gmail.com>